Move the precision/recall/f1 and confusion matrix computations into a separate metrics.py file.
train.py has gotten very large already, and this seems like a decent place to draw a line.
